re #1259: Refactor transform rules into a common super class.
