+++ warehouse/query-core/src/test/java/datawave/query/planner/rules/NodeTransformVisitorTest.java
@@ -11,7 +11,6 @@ import org.apache.commons.jexl2.parser.ASTAndNode;
 import org.apache.commons.jexl2.parser.ASTJexlScript;
 import org.apache.commons.jexl2.parser.JexlNode;
 import org.apache.commons.jexl2.parser.ParseException;
-import org.apache.hadoop.yarn.server.api.protocolrecords.RefreshUserToGroupsMappingsRequest;
 import org.apache.log4j.Logger;
 import org.junit.Before;
 import org.junit.Test;
